+++ R/enrich_method.R
@@ -98,24 +98,24 @@ setMethod("enrich", "context", function(.Object, sAttribute = NULL, pAttribute =
   # .Object2@stat <- copy(.Object@stat)
   if (!is.null(sAttribute)){
     # check that all s-attributes are available
-    if (verbose) message("... checking that all s-attributes are available")
+    .verboseOutput("... checking that all s-attributes are available", verbose)
     stopifnot( all(sAttribute %in% CQI$attributes(.Object@corpus, type = "s")) )
     
     for (sAttr in sAttribute){
-      if (verbose) message("... get struc for s-attribute: ", sAttr)
+      .verboseOutput(sprintf("get struc for s-attribute: %s", sAttr), verbose)
       strucs <- CQI$cpos2struc(.Object@corpus, sAttr, .Object@cpos[["cpos"]])
       if (id2str == FALSE){
         colname_struc <- paste(sAttr, "int", sep = "_")
         if (colname_struc %in% colnames(.Object@cpos)){
-          if (verbose) message("... already present, skipping assignment of column: ", colname_struc)
+          .verboseOutput(sprintf("... already present, skipping assignment of column: %s", colname_struc), verbose)
         } else {
           .Object@cpos[[colname_struc]] <- strucs
         }
       } else {
         if (sAttr %in% colnames(.Object@cpos)){
-          if (verbose) message("... already present, skipping assignment of column: ", sAttr)
+          .verboseOutput(sprintf("... already present, skipping assignment of column: %s", sAttr), verbose)
         } else {
-          if (verbose) message("... get string for s-attribute: ", sAttr)
+          .verboseOutput(sprintf("... get string for s-attribute: %s", sAttr), verbose)
           strings <- CQI$struc2str(.Object@corpus, sAttr, strucs)
           .Object@cpos[[sAttr]] <- as.nativeEnc(strings, from = .Object@encoding)
         }
@@ -124,18 +124,18 @@ setMethod("enrich", "context", function(.Object, sAttribute = NULL, pAttribute =
   }
   if (!is.null(pAttribute)){
     # check that all p-attributes are available
-    if (verbose) message("... checking that all p-attributes are available")
+    .verboseOutput("... checking that all p-attributes are available", verbose)
     stopifnot( all(pAttribute %in% CQI$attributes(.Object@corpus, type = "p")) )
     
     # add ids
     for (pAttr in pAttribute){
       colname <- paste(pAttr, "id", sep = "_")
       if (colname %in% colnames(.Object@cpos)){
-        if (verbose) message("... already present - skip getting ids for p-attribute: ", pAttr)
+        .verboseOutput(sprintf("already present - skip getting ids for p-attribute: %s", pAttr), verbose)
       } else {
-        if (verbose) message("... getting token id for p-attribute: ", pAttr)
+        .verboseOutput(sprintf("getting token id for p-attribute: %s", pAttr), verbose)
         ids <- CQI$cpos2id(.Object@corpus, pAttr, .Object@cpos[["cpos"]])
-        if (verbose) message("... assigning to data.table")
+        .verboseOutput("assigning to data.table", verbose)
         .Object@cpos[[colname]] <- ids
       }
     }
@@ -144,9 +144,9 @@ setMethod("enrich", "context", function(.Object, sAttribute = NULL, pAttribute =
     if (id2str){
       for (pAttr in pAttribute){
         if (pAttr %in% colnames(.Object@cpos)){
-          if (verbose) message("... already present - skip getting strings for p-attribute: ", pAttr)
+          .verboseOutput(sprintf("already present - skip getting strings for p-attribute: %s", pAttr), verbose)
         } else {
-          if (verbose) message("... id2str for p-attribute: ", pAttr)
+          .verboseOutput(sprintf("... id2str for p-attribute: %s", pAttr), verbose)
           decoded <- CQI$id2str(.Object@corpus, pAttr, .Object@cpos[[paste(pAttr, "id", sep = "_")]])
           .Object@cpos[[pAttr]] <- as.nativeEnc(decoded, from = .Object@encoding)
           .Object@cpos[[paste(pAttr, "id", sep = "_")]] <- NULL